+++ sprocs/index.js
@@ -117,6 +117,7 @@ module.exports = {
             'variants_unlink.sql',
             'grader_loads_current.sql',
             'server_loads_current.sql',
+            'server_usage_current.sql',
             'assessment_questions_calculate_stats_for_assessment.sql',
             'assessment_questions_calculate_stats.sql',
             'instance_questions_calculate_stats.sql',